backend/requirements.txt
Django==5.0.1
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
#django-otp==1.3.0
qrcode==7.4.2
requests==2.31.0
social-auth-app-django==5.4.0
python-jose==3.3.0
python-dotenv==1.0.0
# Testing
coverage==6.5.0
freezegun==1.2.2
pyotp==2.9.0
django-otp==1.1.6
whitenoise==6.6.0
djangorestframework-simplejwt==5.3.0